Removing logging.
Removing pretty_env_log as it was having a 10/15% performance impact in
the benchmark which showed no logging.

The logging also seemed to cause the benchmark to change 2/3 randomly up
or down each run. Without the logging the result it now alot more
stable and variaries 0.1% up or down each time to the last mean.